+++ internal/session/stop.go
@@ -2,36 +2,16 @@ package session
 
 import (
 	"context"
-	"time"
-
-	"github.com/chromedp/cdproto/browser"
-	"github.com/chromedp/chromedp"
 
 	"github.com/KaraBala10/chatbang-pro/internal/config"
 )
 
-func stopBrowser(ctx context.Context, ctxCancel, allocCancel context.CancelFunc, profileDir string) {
-	if ctx != nil {
-		closeCtx, cancel := context.WithTimeout(ctx, 2*time.Second)
-		_ = chromedp.Run(closeCtx, chromedp.ActionFunc(func(ctx context.Context) error {
-			return browser.Close().Do(ctx)
-		}))
-		cancel()
+func stopBrowser(_ context.Context, tabCancel, allocCancel context.CancelFunc, profileDir string) {
+	if tabCancel != nil {
+		tabCancel()
 	}
-	if ctxCancel != nil {
-		ctxCancel()
+	if allocCancel != nil {
+		allocCancel()
 	}
 	config.ReleaseProfile(profileDir)
-	if allocCancel == nil {
-		return
-	}
-	done := make(chan struct{})
-	go func() {
-		allocCancel()
-		close(done)
-	}()
-	select {
-	case <-done:
-	case <-time.After(3 * time.Second):
-	}
 }